Create onboarding step pages
